only use force when prompted in the failure message (#325)
* only use force when prompted in the failure message

* update unit tests

* move initialise and terminate block device driver methods to device_plugin action_plugin to be loaded when agent starts in monitored mode

* update unit tests to reflect change in location of adding initialise and terminate start-up and teardown functions

* wipes signatures on lvcreate call to fix efs failures

* Revert "wipes signatures on lvcreate call to fix efs failures"

This reverts commit 0ad8a98a22d62e854646fb52c492a4962970619d.
